fix document generation for createFeature
